requirements-dev.txt
# Dependências para desenvolvimento

# Aplicação
streamlit==1.28.1
llama-parse==0.1.1
python-dotenv==1.0.0

# Desenvolvimento e Testes
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
pytest==7.4.3
pytest-cov==4.1.0

# Documentação
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Qualidade de Código
autopep8==2.0.4
isort==5.13.2

# Utilitários
ipython==8.18.1
jupyter==1.0.0
